# Faster map with mulitple arguments