Hi Jos,

This makes it faster to map multiple arrays or matrices (with broadcasting), by not creating the internally broadcasted matrices unless the callback specifically needs it.

About 4x faster when no index is needed, 3x faster when an index is needed and is the same when also the broadcasted arrays or matrices are needed.

<img width="872" height="455" alt="image" src="https://github.com/user-attachments/assets/b29a5807-7fae-4446-8503-04c81445b3ba" />

develop

```
abs(genericMatrix)                               1.41 µs   ±0.13%
abs(array)                                       1.28 µs   ±0.24%
abs(numberMatrix)                                1.42 µs   ±0.22%
genericMatrix.map(abs)                           5.01 µs   ±0.19%
numberMatrix.map(abs)                            5.01 µs   ±0.17%
map(genericMatrix, abs)                          5.03 µs   ±0.19%
map(numberMatrix, abs)                           5.01 µs   ±0.19%
map(array, abs)                                  5.01 µs   ±0.20%
map(array, abs.signatures.number)                2.23 µs   ±0.26%
genericMatrix.map(abs.signatures.number)         1.80 µs   ±5.91%
numberMatrix.map(abs.signatures.number)          1.70 µs   ±0.24%
genericMatrix iterate                            7.46 µs   ±0.25%
map(rowMatrix, columnMatrix, multiCallback)     35.78 µs   ±0.25%
map(rowArray, columnArray, multiCallback)       29.32 µs   ±0.20%
map(rowMatrix, columnMatrix, multiCallback1)    40.34 µs   ±0.19%
map(rowArray, columnArray, multiCallback1)      34.08 µs   ±0.19%
map(rowMatrix, columnMatrix, multiCallback2)    40.48 µs   ±0.17%
map(rowArray, columnArray, multiCallback2)      34.28 µs   ±0.20%
```

this PR
```
abs(genericMatrix)                               1.43 µs   ±0.11%
abs(array)                                       1.28 µs   ±0.25%
abs(numberMatrix)                                1.45 µs   ±1.80%
genericMatrix.map(abs)                           5.16 µs   ±0.94%
numberMatrix.map(abs)                            5.13 µs   ±0.30%
map(genericMatrix, abs)                          5.11 µs   ±0.18%
map(numberMatrix, abs)                           5.09 µs   ±0.24%
map(array, abs)                                  5.01 µs   ±0.17%
map(array, abs.signatures.number)                2.22 µs   ±0.29%
genericMatrix.map(abs.signatures.number)         1.71 µs   ±0.27%
numberMatrix.map(abs.signatures.number)          1.72 µs   ±0.62%
genericMatrix iterate                           12.31 µs   ±6.69%
map(rowMatrix, columnMatrix, multiCallback)      6.89 µs   ±0.25%
map(rowArray, columnArray, multiCallback)        6.75 µs   ±0.22%
map(rowMatrix, columnMatrix, multiCallback1)    12.42 µs   ±0.23%
map(rowArray, columnArray, multiCallback1)      12.28 µs   ±0.21%
map(rowMatrix, columnMatrix, multiCallback2)    39.67 µs   ±0.21%
map(rowArray, columnArray, multiCallback2)      32.78 µs   ±0.20%
```